commit created by maybePromptToGcam()
