+++ src/aws/lock.py
@@ -6,7 +6,7 @@ from python_dynamodb_lock.python_dynamodb_lock import DynamoDBLockClient  # type
 
 dynamodb_resource = boto3.resource("dynamodb", region_name=AWS_REGION)
 
-lock_client = DynamoDBLockClient(
+dynamodb_lock_client = DynamoDBLockClient(
     dynamodb_resource,
     table_name=LOCK_TABLE,
     lease_duration=timedelta(seconds=20),
