+++ test/asana/helpers/test_asana_comment_from_github_comment.py
@@ -1,5 +1,6 @@
 from html import escape
 from unittest.mock import patch
+
 from src.asana.helpers import asana_comment_from_github_comment
 from test.impl.mock_dynamodb_test_case import MockDynamoDbTestCase
 from test.impl.builders import builder, build
@@ -7,7 +8,7 @@ from test.test_utils import magic_mock_with_return_type_value
 
 
 @patch(
-    "src.dynamodb.client.get_asana_domain_user_id_from_github_handle",
+    "src.aws.s3_client.get_asana_domain_user_id_from_github_handle",
     magic_mock_with_return_type_value(
         {"github_test_user_login": "TEST_USER_ASANA_DOMAIN_USER_ID"}
     ),
